requirements.txt
# requirements.txt
google-generativeai==0.8.2
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.15.0
openai==1.40.0
moviepy==1.0.3
whisper==1.1.10
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
fake-useragent==1.4.0
Pillow==10.1.0
